rename: 회원가입 로직 변경으로 인한 폴더 이동
