+++ src/auth/components/signup/CreatePassword.tsx
@@ -90,7 +90,7 @@ const CreatePassword = ({ onNext }: { onNext: (password: string) => void }) => {
           variant={isFormValid ? 'default' : 'secondary'}
           disabled={!isFormValid} // 버튼 비활성화 적용
         >
-          다음으로
+          회원가입 완료
         </AuthButton>
       </div>
     </div>
