+++ src/user/components/mentor/MentorProfile.tsx
@@ -1,16 +1,17 @@
 import { useState, useEffect } from 'react';
 import { AuthButton } from '@/auth/components/AuthButton';
 import { AuthInputBox } from '@/auth/components/AuthInputBox';
-import ModalInput from '@/auth/components/ModalInput';
+import ModalInput from '@/user/components/ModalInput';
 
-const MenteeProfile = ({ onNext }: { onNext: () => void }) => {
+const MentorProfile = ({ onNext }: { onNext: () => void }) => {
   const [bio, setBio] = useState(''); // 자기소개 상태
   const [nickname, setNickname] = useState(''); // 닉네임 상태
   const [isNicknameValid, setIsNicknameValid] = useState(false); // 닉네임 유효성 상태
   const [nicknameMessage, setNicknameMessage] = useState(''); // 닉네임 메시지
   const [profileImage, setProfileImage] = useState<string | null>(null); // 프로필 이미지 상태
   const [selectedExperience, setSelectedExperience] = useState(''); // 경력 선택 상태
-  const [major, setMajor] = useState(''); // 회사 입력 상태
+  const [company, setCompany] = useState(''); // 회사 입력 상태
+  const [position, setPosition] = useState(''); // 직무 입력 상태
 
   const [isFormValid, setIsFormValid] = useState(false); // 전체 폼 유효성
 
@@ -45,10 +46,11 @@ const MenteeProfile = ({ onNext }: { onNext: () => void }) => {
   useEffect(() => {
     const isFormFilled =
       isNicknameValid &&
-      major.trim().length > 0 &&
+      company.trim().length > 0 &&
+      position.trim().length > 0 &&
       selectedExperience.trim().length > 0;
     setIsFormValid(isFormFilled);
-  }, [isNicknameValid, major, selectedExperience]);
+  }, [isNicknameValid, company, position, selectedExperience]);
 
   return (
     <div className="flex flex-col w-full h-[calc(100vh-15vh)] overflow-hidden">
@@ -113,23 +115,46 @@ const MenteeProfile = ({ onNext }: { onNext: () => void }) => {
           />
         </div>
 
-        {/* 전공 입력 */}
+        {/* 회사 입력 */}
         <div className="mt-6">
           <AuthInputBox
-            label="전공 분야"
+            label="회사"
             type="text"
-            placeholder="분야를 입력해 주세요."
-            value={major}
-            onChange={(e) => setMajor(e.target.value)}
+            placeholder="현재 재직 중인 회사 명을 입력해주세요."
+            value={company}
+            onChange={(e) => setCompany(e.target.value)}
           />
         </div>
 
-        {/* 학력 입력 (ModalInput 사용) */}
+        {/* 직무 입력 */}
+        <div className="mt-6">
+          <AuthInputBox
+            label="직무"
+            type="text"
+            placeholder="직무를 입력해 주세요."
+            value={position}
+            onChange={(e) => setPosition(e.target.value)}
+          />
+        </div>
+
+        {/* 경력 입력 (ModalInput 사용) */}
         <div className="mt-6">
           <ModalInput
-            label="학력"
-            placeholder="학력을 선택해 주세요."
-            options={['재학 중', '석사', '박사', '학사', '기타']}
+            label="경력"
+            placeholder="총 경력을 선택해 주세요."
+            options={[
+              '1년',
+              '2년',
+              '3년',
+              '4년',
+              '5년',
+              '6년',
+              '7년',
+              '8년',
+              '9년',
+              '10년',
+              '11년 이상',
+            ]}
             value={selectedExperience}
             onChange={setSelectedExperience}
           />
@@ -146,9 +171,9 @@ const MenteeProfile = ({ onNext }: { onNext: () => void }) => {
             style={{ whiteSpace: 'pre-wrap' }}
             placeholder={`예시)
 안녕하세요:)
-저는 00대학교 경영학과 2학년에 재학중인 ‘바이’입니다.
+저는 00대학교 경영학과를 졸업하고 현재 xxxx에 다니고 있는 ‘바이’입니다.
 
-제가 꿈꾸는 분야는 마케팅 분야로, 다양한 산업에서 성공적인 마케팅 캠페인을 경험해보고 싶습니다.`}
+저는 브랜드 마케팅 전문가로, 다년간 다양한 산업에서 성공적인 마케팅 캠페인을 기획하고 실행해온 경험을 보유하고 있습니다. 특히, 데이터를 분석하여 전략을 수립하고 소비자 인사이트를 반영한 솔루션 개발에 강점이 있습니다.`}
             value={bio}
             onChange={handleBioChange}
           />
@@ -171,4 +196,4 @@ const MenteeProfile = ({ onNext }: { onNext: () => void }) => {
   );
 };
 
-export default MenteeProfile;
+export default MentorProfile;
